Improve java 17 switch expression tests
