+++ src/main/resources/META-INF/rewrite/java-version-21.yml
@@ -136,7 +136,7 @@ recipeList:
 type: specs.openrewrite.org/v1beta/recipe
 name: org.openrewrite.java.migrate.SwitchPatternMatching
 displayName: Adopt switch pattern matching (JEP 441)
-description: ->-
+description: >-
   [JEP 441](https://openjdk.org/jeps/441) describes how some switch statements can be improved with pattern matching.
   This recipe applies some of those improvements where applicable.
 tags:
